refactor : 입력부와 게임 시작 부분 분리
